Warning fixes in yuni-docmake
git-svn-id: http://svn.ta3d.org/yuni/trunk@1747 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/tools/yuni-docmake/indexes.cpp
@@ -639,7 +639,7 @@ namespace DocIndex
 				if (!termID || !scount)
 					continue;
 				unsigned int count = scount.to<unsigned int>();
-				double d = 1. - (count * 0.5 / maxO);
+				double d = 1. - (count * 0.5 / (double)maxO);
 				s << "UPDATE terms SET weight_rel_others = " << d << " WHERE id = " << termID << ";\n";
 			}
 		}
